This is a big commit, and it will break your build unless you define NCL_INSTALL_DIR to point to the installation directory of NCL 2.1 (the directory that you used when you passed in the --prefix arg to configure for NCL).
Paul, I'm afraid that you'll have to figure out how to do this on Windows.
I had to add a -I and a -L flag to the Jamroot for Mac.

Dave, you may have to rename $NCL_INSTALL_DIR/lib/libncl-2.1.02 to $NCL_INSTALL_DIR/lib/libncl.so 
to get bjam to recognize it.

In testing, I introduced thin wrappers around tree descriptions and data matrices.

I also moved to a system in which a tree "builder" is used to construct the tree

the Newick class in phycas/__init__.py is one example of a builder.

The other is the TreeDescription class in _NexusReader.py

Those changes came about because the CIPRES tree description returned 0-based newick
strings, while the NCL 2.1 version returns 1-based strings.




git-svn-id: https://phycas.svn.sourceforge.net/svnroot/phycas/trunk@688 b774380d-c90d-0410-997c-aec342e5d6fd