Reduced stringency of Relative Rate check
Reduced tolerance used to detect when the elements of a transformed Relative Rate distribution (equivalent to a Dirichlet distribution) do not add up to 1.0. Formerly, the sum needed to be within 0.0000001 of 1.0 to be acceptable; now they need only be within 0.001. This was done to allow such distributions to be created from reference distribution descriptions, where the precision of the parameters supplied end up yielding not-so-precise transformed relative rate distributions.